+++ test/resources/language/rust/cargo-targets/examples/a.rs
@@ -1,0 +1,4 @@
+fn main() { let foo_ex_a = 0; }
+
+#[test]
+fn test() { let foo_ex_a_test = 0; }