+++ test/resources/language/rust/cargo-targets/src/a.rs
@@ -1,0 +1,4 @@
+fn foo_a() {}
+
+#[test]
+fn test() { let foo_a_test = 0; }