+++ test/resources/language/rust/cargo-targets/src/bin/a.rs
@@ -1,0 +1,4 @@
+fn main() { let foo_bin_a = 0; }
+
+#[test]
+fn test() { let foo_bin_a_test = 0; }