+++ test/resources/language/rust/cargo-targets/src/lib.rs
@@ -1,0 +1,6 @@
+mod a;
+
+fn foo_lib() {}
+
+#[test]
+fn test() { let foo_lib_test = 0; }