+++ test/resources/language/rust/cargo-targets/src/main.rs
@@ -1,0 +1,4 @@
+fn main() { let foo_main = 0; }
+
+#[test]
+fn test() { let foo_main_test = 0; }